adding the colormaps branch - colormaps implemented following ncview - added this to the mesh class but it should be moved   into its own class and the mesh class should inherit colormaps - currently hardcoded in mesh and colormap constructs for   sea surface temperature and sea height data sets. this should   be generalized back into the netcdf class - no room to store big data sets in git. you need:   * bathymetry dataset   * the netcdf data sets   * global high res coastline datasets (optional - i have turned these off) - no texture mapping so no textures are required - modifies the passthrough fragment shader to adjust the lighting - need to rename the passthrough shader - lighting shader code can be simplified (remove branches) - current branch setup is for the artmap project, but can merge back to head   after some redesign

+++ include/netcdfIO.h
@@ -19,11 +19,11 @@ class ncio{
         string  lat_name;
         string  lon_name;
         string  field_name;
-    
+
         size_t nlat;
         size_t nlon;
-        std::vector<float>  lat;
-        std::vector<float>  lon;
+        std::vector<double>  lat;
+        std::vector<double>  lon;
         //std::vector< std::vector<float> >  field;
         std::vector<float> field;
